#23 constraints to project

+++ src/main/kotlin/bg/elsys/jobche/service/TaskService.kt
@@ -1,23 +1,31 @@
 package bg.elsys.jobche.service
 
+import bg.elsys.jobche.config.security.AuthenticationDetails
 import bg.elsys.jobche.entity.body.task.TaskBody
 import bg.elsys.jobche.entity.model.Task
-import bg.elsys.jobche.entity.response.TaskPaginatedResponse
 import bg.elsys.jobche.entity.response.TaskResponse
+import bg.elsys.jobche.exceptions.TaskModificationForbiddenException
 import bg.elsys.jobche.exceptions.TaskNotFoundException
 import bg.elsys.jobche.repositories.TaskRepository
+import bg.elsys.jobche.repositories.UserRepository
 import org.springframework.data.domain.PageRequest
 import org.springframework.data.domain.Pageable
 import org.springframework.stereotype.Service
 
 @Service
-class TaskService(val taskRepository: TaskRepository) {
+class TaskService(val taskRepository: TaskRepository,
+                  val userRepository: UserRepository,
+                  val authenticationDetails: AuthenticationDetails) {
     fun create(taskBody: TaskBody): TaskResponse {
+        //Get Id of the creator
+        val user = userRepository.findByEmail(authenticationDetails.getEmail())
+
         val task = taskRepository.save(Task(taskBody.title,
                 taskBody.description,
                 taskBody.payment,
                 taskBody.numberOfWorkers,
-                taskBody.dateTime))
+                taskBody.dateTime,
+                user!!.id))
 
         return TaskResponse(task.id,
                 task.title,
@@ -40,19 +48,35 @@ class TaskService(val taskRepository: TaskRepository) {
     }
 
     fun update(task: TaskBody, id: Long) {
+        //Get the id of the user that is requesting the update
+        val user = userRepository.findByEmail(authenticationDetails.getEmail())
+
         if(taskRepository.existsById(id)) {
             val taskToUpdate = taskRepository.getOne(id)
+
+            if(taskToUpdate.creatorId != user?.id) {
+                throw TaskModificationForbiddenException()
+            }
+
             taskToUpdate.title = task.title
             taskToUpdate.payment = task.payment
             taskToUpdate.numberOfWorkers = task.numberOfWorkers
             taskToUpdate.description = task.description
             taskToUpdate.dateTime = task.dateTime
             taskRepository.save(taskToUpdate)
+
         } else throw TaskNotFoundException()
     }
 
     fun delete(id: Long) {
         if(taskRepository.existsById(id)) {
+            val task = taskRepository.findById(id)
+            val user = userRepository.findByEmail(authenticationDetails.getEmail())
+
+            if(task.get().creatorId != user!!.id) {
+                throw TaskModificationForbiddenException()
+            }
+
             taskRepository.deleteById(id)
         } else throw TaskNotFoundException()
     }
